requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-dotenv>=1.0.0
httpx>=0.25.0
google-generativeai>=0.3.0
loguru>=0.7.2
python-dateutil>=2.8.2
python-multipart>=0.0.6
pydantic-settings>=2.0.0

# LangWatch and Observability
langwatch>=0.1.0
openinference-instrumentation-google-adk>=0.1.0

# Google ADK
google-adk>=0.2.0
google-cloud-aiplatform>=1.38.0

# Optional but recommended
litellm>=1.0.0
